Add Like in post feature
Add "Like" feature with API, service, and repository

Introduced a new "Like" feature to allow users to like/unlike posts.
Updated the project structure to include folders for "Like" handlers
and models. Registered `ILikeRepository` and `ILikeSerivce` for
dependency injection.

Implemented `SetLikeCommandHandler` to process "Set Like" commands
using MediatR. Added `SetLikeCommand` to define the command structure.
Created `ILikeRepository` and `LikeRepository` for database operations
and `ILikeSerivce` and `LikeService` for business logic.

Added `LikeDto` for data transfer and `LikeController` to expose an
API endpoint for the "Set Like" functionality. Ensured modularity and
maintainability in the implementation.